Struct ogma::eng::Block

pub struct Block<'a> {
    pub in_ty: Type,
    /* private fields */
}

A compilation unit for a single ast::Block.

The block is foundational to the compilation engine. It acts as the interface between the engine’s internal mechanisms and the implementers, exposing an API which is ergonomic, while handling all the typing and checking behind the scenes.

Fields
in_ty: Type

The block’s input type.

Implementations

impl<'a> Block<'a>

pub fn next_arg(&mut self) -> Result<Box<ArgBuilder<'_>>, Error>

Get the Block’s next argument.

The argument is agnostic to whether it is a variable, literal, or expression. The return type is an argument builder, which can be used to assert type information about the argument. Once the assertations are done, use .concrete() to resolve that the types are known and an [Argument] is produced.

impl<'a> Block<'a>

pub fn op_tag(&self) -> &Tag

The operation (command) tag.

pub fn blk_tag(&self) -> &Tag

The entire block tag.

pub fn in_ty(&self) -> &Type

The input Type of the block.

pub fn args_len(&self) -> usize

Returns the number of arguments remaining.

pub fn defs(&self) -> &'a Definitions

The current definitions being compiled with.

pub fn compiler(&self) -> &'a Compiler<'_>

A Compiler reference.

pub fn assert_input(&self, ty: &Type) -> Result<(), Error>

Assert the block as the given input type, constructing an error if not.

pub fn assert_output(&mut self, ty: Type)

Assert that this block will return the given type.

Asserting an output type gives the type inferer knowledge about this block’s output. It is recommended to assert this as soon as possible in the implementation, so if compilation fails, the output information is still captured.

Panics

For debug builds, panics if an output type has already been set on this block. Only a single assertion should be made in an implementation.

For debug builds, this assertion is checked against the evalulation output type.

pub fn insert_anon_type_into_compiler(&mut self, ty: Type)

Inserts a new, anonymous type into the compiler.

Note that this will not affect already resolved nodes, only inferred nodes.

pub fn get_flag<'b, N: Into<Option<&'b str>>>(&mut self, flag: N) -> Option<Tag>

Gets the flag that matches a given name.

If no name is given with None, the first flag first is returned, if there is one.

The flag is removed from the flag stack.

pub fn peek_next_arg_node(&self) -> Option<ArgNode>

See if there is a next argument node, without popping off the stack.

pub fn peek_last_arg_node(&self) -> Option<ArgNode>

See if there is a final argument node, without popping off the stack.

pub fn assert_adds_vars(&mut self, retain_op_coupling: bool)

Assert that this operation will be adding variables.

This should be called early on to flag to the compiler that it cannot eagerly seal this node. Usually, retain_op_coupling: false, unless one or more arguments cannot be considered sealed until this operation is (see the fold intrinsic as an example). If retain_op_coupling: true, any arguments that do not require the op to be considered sealed should flag this with [ArgBuilder::decouple_op_seal].

Note that this function should be paired with a subsequent Block::assert_vars_added call once all required variables have been added.

pub fn assert_vars_added(&mut self)

Asserts that all the variables have been added, allowing the compiler to seal this node.

This should be paired with a Block::assert_adds_vars call, and invoked after the variables have been added.

pub fn create_var_ref(
&mut self,
arg: ArgNode,
ty: Type
) -> Result<Variable, Error>

Assert this argument is a variable and construct a reference to it.

If the block does not contain a reference to an up-to-date locals, and error is returned.

Note that the variable will be defined at the next block in the chain, not the current one.

Type safety

The variable will be created expecting the type ty. set_data only validates types in debug builds, be sure that testing occurs of code path to avoid UB in release.

pub fn inject_manual_var_into_arg_locals<N>(
&mut self,
arg: ArgNode,
name: N,
ty: Type
) -> Result<Variable, Error> where
N: Into<Str>,

Create a variable reference not off a specific argument, but by manually specifying the variable name.

This is useful for expressions that need to supply more than just the input. For instance the fold command will supply the $row variable which is a track of the TableRow.

Consider using the helper Block::inject_manual_var_next_arg which uses the next arguments argnode.

pub fn inject_manual_var_next_arg<N>(
&mut self,
name: N,
ty: Type
) -> Result<Variable, Error> where
N: Into<Str>,

Helper for Block::inject_manual_var_into_arg_locals which peeks the next argument on the stack to use as the arg input.

pub fn oblige_args_supplied_tys<T: Into<Option<Type>>>(&mut self, ty: T)

Oblige the remaining arguments to have the input type ty.

ty is similar to [ArgBuilder::supplied], where a None indicates to use the block’s input type.

This method does not pop any arguments, it merely flags to the compiler to insert type graph changes.

impl<'a> Block<'a>

Evalulation functions.

pub fn eval<F>(self, out_ty: Type, f: F) -> Result<Step, Error> where
F: Fn(Value, Context<'_>) -> Result<(Value, Environment), Error>,
F: Func<Result<(Value, Environment), Error>>,

Most flexible evaluation option, but also most brittle.

BE EXTRA CAREFUL WITH THE out_ty THAT IT MATCHES THE EVAL VALUE. It is recommended to thoroughly test code paths through here to ensure type validity.

Usage

Input value (Value) should be cast to expected input type (use .try_into()?).

Arguments can use this type if they are expecting the blocks input.

Arguments

Arguments do not need to use blocks input. If no input type is needed, the argument can be built with Type::Nil and Value::Nil can be passed on through!

pub fn eval_o<F, O>(self, f: F) -> Result<Step, Error> where
F: Fn(Value, Context<'_>) -> Result<(O, Environment), Error>,
F: Func<Result<(O, Environment), Error>>,
O: AsType + Into<Value>,

Preferred way of creating a eval step.

This supplies the Value input but uses type inference on O to get the output type.

impl<'a> Block<'a>

pub fn output_ty(&mut self) -> Option<Type>

Looks into the TG and fetches if this block’s output type is known. It also will infer the output type by deduction from TG context.

If None type is going to be returned, flag the block’s node as one that might have to have its output inferred.