Module ogma::lang

Language characteristics.

Modules

ast

AST representation.

parse

Parsing source into a AST.

Structs

Definitions

The working set of ogma definitions.

ImplEntry

An implementation entry.

OgmaData

A non-primitive data structure.

Table

Wrapper of a table, used to specialise drop and cloning behaviour.

Enums

Type

Ogma data types.

Value

The value of an evaluation.

Traits

AsType

This Rust type represents an ogma type.

Functions

construct_def_table

Construct a table comprising of the various definitions in defs.

process_definition

Process an expression whichs defines a definition.

recognise_definition

Recognise a string which defines an implementation or type.