Module ogma::lang::ast

AST representation.

This is comprised of:

in data.csv | each-row { max | + 1 } | echo
   ^------^            ^-----------^        --> Args
^---------^   ^--------------------^   ^--^ --> Blocks
^-----------------------------------------^ --> Expression

Structs

BlockParts

The concrete parts of a block.

DefinitionImpl

A implementation definition.

DefinitionType

A type definition.

DotOperatorBlock

Dot operator has terms that are [lhs,rhs].

Expression

An expression. An expression is made of multiple blocks connected by pipes.

Field

A product type field.

Parameter

A implementation definition’s parameters.

PrefixBlock

Polish notation block: op term1 term2.

Tag

Tag only implements tag value equality checking (as in .str()).

Tag_

Inner tag.

Variant

A sum type variant.

Enums

Argument

An argument kind.

Location

A location indicator.

Term

A expression block term: bar in foo bar.

TypeVariant

Types are either Sum (enum) or Product (struct).

Traits

IBlock

The common aspects of a processing block.

Type Definitions

Block

A single expression block. An expression is made of multiple blocks connected by pipes.

Op

Operation tag.

Terms

List of terms.