Trait ogma::lang::ast::IBlock

pub trait IBlock: Debug + Send + Sync {
    fn op(&self) -> Cow<'_, Tag>;
    fn terms(&self) -> Cow<'_, [Term]>;
    fn in_ty(&self) -> Option<Cow<'_, Tag>>;
    fn out_ty(&self) -> Option<Cow<'_, Tag>>;
    fn block_tag(&self) -> Tag;
    fn parts(self: Box<Self>) -> BlockParts;
    fn clone(&self) -> Block;
}

The common aspects of a processing block.

Blocks can be in different notation style, but at their core always specify a single op tag, a list of terms, optional input type constraint, and optional output type constraint.

Required Methods

fn op(&self) -> Cow<'_, Tag>

The operation.

fn terms(&self) -> Cow<'_, [Term]>

The terms.

fn in_ty(&self) -> Option<Cow<'_, Tag>>

An optional annotated input type constraint.

fn out_ty(&self) -> Option<Cow<'_, Tag>>

An optional annotated ouput type constraint.

fn block_tag(&self) -> Tag

Create a tag that extends across the whole block.

fn parts(self: Box<Self>) -> BlockParts

The common aspects of a a block: the operation tag and a list of terms.

fn clone(&self) -> Block

Trait-object safe clone.