Function ogma::lang::process_definition

pub fn process_definition<'a>(
    def: &'a str, 
    loc: Location, 
    help: Option<String>, 
    defs: &mut Definitions
) -> Result<(Value, Option<&'a str>), Error>

Process an expression whichs defines a definition.

If successful the definition will be added to defs.

Return Value

Sometimes a definition will process and return a Value, but will have trailing ogma code that potentially might compile to use the the returned Value. This is why the successful return value is Ok((Value, Option<&str>)). The optional returned slice indicates that this slice should be sent through the ogma processor again, with the returned Value as the input. It does not indicate whether it will succeed or not.