Struct ogma::lang::Table

pub struct Table(pub Arc<Table<Value>>);

Wrapper of a table, used to specialise drop and cloning behaviour.

Tuple Fields
0: Arc<Table<Value>>
Implementations

impl Table

pub fn get_mut(&mut self) -> Option<&mut Table<Value>>

If this table reference is unique, get a mutable alias to it.

pub fn make_mut(&mut self) -> &mut Table<Value>

Make a unique reference to this table by cloning if required.