Module ogma::rt::bat

Batch processing of definitions and expressions.

This module provides some structure around batch processing, supplying reporting structures and evaluation semantics.

Structs

Batch

A batch to process.

BatchItem

An item to process.

BatchItemProgress

Serialisable progress of a BatchItem.

Enums

ItemType

The type of this item.

Outcome

The outcome of an item.

OutcomeProgress

The serialisable version of Outcome.

Functions

parse_file

Parse the contents of a file into a list of BatchItems.

parse_str

Parse a string of items.

process

Process a Batch.

Type Definitions

BatchProgress

Serialisable progress of a Batch.