Function ogma::rt::bat::process

pub fn process(
    batch: &Batch, 
    root: &Path, 
    wd: &Path, 
    progress: &ProgressTx, 
    definitions: Definitions
) -> Vec<(Outcome, Duration)>

Process a Batch.

Processing will process each item, doing definitions in a sequential order, and expressions in either parallel or sequential order based on the parallelise field. If fail_fast is true, the first time an error is encountered, subsequent processing will halt. Items that do not get processed will have an outcome of Outstanding. Each process call uses a passed in Definitions which it owns meaning it should not interfere with REPL or other batch instances (store caching is still shared).

Cancellation

The cancellation switch found in cx is checked before each item is processed. It is not checked whilst processing an item.

Progress Reporting

Upon finishing processing an item, a progress report is sent (through cx.progress: ProgressTx::send). This report serialises BatchProgress, which is a status report of each item to be processed. BatchProgress serialises to json, and the processor on the other side should deserialise json back to a BatchProgress to provide detailed progress reporting.