Function ogma::rt::process_expression

pub fn process_expression<I, S>(
    seed: I,
    expr: S,
    loc: Location,
    defs: &Definitions,
    root: &Path,
    wd: &Path
) -> Result<Value, Error>where
    I: AsType + Into<Value> + 'static,
    S: Into<Arc<str>>,

Parse and evaluate an expr, returning the value if successful.

root: The root directory that the ogma instance is evaluating in. wd: The working directory that the expression is evaluating in.

These two paths are important since commands such as open and ls are relative. There are also security concerns with accessing items above the root path, so this is generally disallowed.